# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Feloxi API",
    "description": "Celery task monitoring and management API",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/v1/alerts/history": {
      "get": {
        "tags": [
          "alerts"
        ],
        "operationId": "list_history",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertHistoryResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/rules": {
      "get": {
        "tags": [
          "alerts"
        ],
        "operationId": "list_rules",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRulesResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "alerts"
        ],
        "operationId": "create_rule",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlertRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRule"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/rules/{rule_id}": {
      "get": {
        "tags": [
          "alerts"
        ],
        "operationId": "get_rule",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "description": "Alert rule ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRule"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "alerts"
        ],
        "operationId": "update_rule",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "description": "Alert rule ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlertRuleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertRule"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "alerts"
        ],
        "operationId": "delete_rule",
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "description": "Alert rule ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlertChannel": {
        "oneOf": [
          {
            "type": "object",
            "required": [
              "webhook_url",
              "type"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "slack"
                ]
              },
              "webhook_url": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "to",
              "type"
            ],
            "properties": {
              "to": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "type": {
                "type": "string",
                "enum": [
                  "email"
                ]
              }
            }
          },
          {
            "type": "object",
            "required": [
              "url",
              "type"
            ],
            "properties": {
              "headers": {
                "type": [
                  "object",
                  "null"
                ],
                "additionalProperties": {
                  "type": "string"
                },
                "propertyNames": {
                  "type": "string"
                }
              },
              "type": {
                "type": "string",
                "enum": [
                  "webhook"
                ]
              },
              "url": {
                "type": "string"
              }
            }
          },
          {
            "type": "object",
            "required": [
              "routing_key",
              "type"
            ],
            "properties": {
              "routing_key": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "pagerduty"
                ]
              }
            }
          }
        ],
        "description": "Alert notification channel configuration."
      },
      "AlertCondition": {
        "oneOf": [
          {
            "type": "object",
            "description": "Trigger when task failure rate exceeds threshold.",
            "required": [
              "threshold",
              "window_minutes",
              "task_name",
              "type"
            ],
            "properties": {
              "task_name": {
                "type": "string"
              },
              "threshold": {
                "type": "number",
                "format": "double"
              },
              "type": {
                "type": "string",
                "enum": [
                  "task_failure_rate"
                ]
              },
              "window_minutes": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              }
            }
          },
          {
            "type": "object",
            "description": "Trigger when queue depth exceeds threshold.",
            "required": [
              "threshold",
              "queue",
              "type"
            ],
            "properties": {
              "queue": {
                "type": "string"
              },
              "threshold": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "type": {
                "type": "string",
                "enum": [
                  "queue_depth"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Trigger when a worker goes offline.",
            "required": [
              "grace_period_seconds",
              "type"
            ],
            "properties": {
              "grace_period_seconds": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "type": {
                "type": "string",
                "enum": [
                  "worker_offline"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Trigger when task duration exceeds threshold at given percentile.",
            "required": [
              "threshold_seconds",
              "percentile",
              "task_name",
              "type"
            ],
            "properties": {
              "percentile": {
                "type": "number",
                "format": "double"
              },
              "task_name": {
                "type": "string"
              },
              "threshold_seconds": {
                "type": "number",
                "format": "double"
              },
              "type": {
                "type": "string",
                "enum": [
                  "task_duration"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Trigger when a beat schedule misses its expected run.",
            "required": [
              "schedule_name",
              "type"
            ],
            "properties": {
              "schedule_name": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "beat_missed"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Trigger when a specific task name fails.",
            "required": [
              "task_name",
              "type"
            ],
            "properties": {
              "task_name": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "task_failed"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Trigger when no events received for a period.",
            "required": [
              "silence_minutes",
              "type"
            ],
            "properties": {
              "silence_minutes": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "type": {
                "type": "string",
                "enum": [
                  "no_events"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Trigger when task throughput deviates from historical baseline (Z-score).",
            "required": [
              "zscore_threshold",
              "window_minutes",
              "task_name",
              "type"
            ],
            "properties": {
              "task_name": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "throughput_anomaly"
                ]
              },
              "window_minutes": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "zscore_threshold": {
                "type": "number",
                "format": "double"
              }
            }
          },
          {
            "type": "object",
            "description": "Trigger when task latency deviates from historical baseline (Z-score).",
            "required": [
              "zscore_threshold",
              "window_minutes",
              "task_name",
              "type"
            ],
            "properties": {
              "task_name": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "latency_anomaly"
                ]
              },
              "window_minutes": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "zscore_threshold": {
                "type": "number",
                "format": "double"
              }
            }
          },
          {
            "type": "object",
            "description": "Trigger when error rate spikes compared to recent baseline.",
            "required": [
              "spike_factor",
              "baseline_hours",
              "task_name",
              "type"
            ],
            "properties": {
              "baseline_hours": {
                "type": "integer",
                "format": "int64",
                "minimum": 0
              },
              "spike_factor": {
                "type": "number",
                "format": "double"
              },
              "task_name": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": [
                  "error_rate_spike"
                ]
              }
            }
          }
        ],
        "description": "Alert condition types."
      },
...
      "AlertHistoryResponse": {
        "type": "object",
        "required": [
          "data",
          "has_more"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlertHistoryRow"
            }
          },
          "has_more": {
            "type": "boolean"
          },
          "total": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
      "AlertHistoryRow": {
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "rule_id",
          "fired_at",
          "severity",
          "summary",
          "details",
          "channels_sent"
        ],
        "properties": {
          "channels_sent": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/ChannelDeliveryStatus"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "details": {
            "type": "object",
            "additionalProperties": {},
            "propertyNames": {
              "type": "string"
            }
          },
          "fired_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "resolved_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "rule_id": {
            "type": "string",
            "format": "uuid"
          },
          "severity": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "AlertRule": {
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "name",
          "is_enabled",
          "condition",
          "channels",
          "cooldown_secs",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "channels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlertChannel"
            }
          },
          "condition": {
            "$ref": "#/components/schemas/AlertCondition"
          },
          "cooldown_secs": {
            "type": "integer",
            "format": "int32"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "is_enabled": {
            "type": "boolean"
          },
          "last_fired_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "name": {
            "type": "string"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AlertRulesResponse": {
        "type": "object",
        "description": "List of alert rules.",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlertRule"
            }
          }
        }
      },
...
      "ChannelDeliveryStatus": {
        "type": "object",
        "description": "Delivery status for a single notification channel.",
        "required": [
          "success"
        ],
        "properties": {
          "error": {
            "type": [
              "string",
              "null"
            ]
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "CreateAlertRuleRequest": {
        "type": "object",
        "required": [
          "name",
          "condition",
          "channels"
        ],
        "properties": {
          "channels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlertChannel"
            }
          },
          "condition": {
            "$ref": "#/components/schemas/AlertCondition"
          },
          "cooldown_secs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "UpdateAlertRuleRequest": {
        "type": "object",
        "required": [
          "name",
          "condition",
          "channels",
          "cooldown_secs",
          "is_enabled"
        ],
        "properties": {
          "channels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AlertChannel"
            }
          },
          "condition": {
            "$ref": "#/components/schemas/AlertCondition"
          },
          "cooldown_secs": {
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_enabled": {
            "type": "boolean"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}